sensor: adxl345: change type of member fifo_samples
The FIFO sample count is stored within a portion of an 8-bit register.
Since only 8-bit configuration registers are involved, it's better to use
uint8_t to avoid unnecessary mixing with uint16_t.

+++ drivers/sensor/adi/adxl345/adxl345.h
@@ -151,7 +151,7 @@ enum adxl345_fifo_mode {
 struct adxl345_fifo_config {
 	enum adxl345_fifo_mode fifo_mode;
 	enum adxl345_fifo_trigger fifo_trigger;
-	uint16_t fifo_samples;
+	uint8_t fifo_samples;
 };
 
 enum adxl345_op_mode {
